requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
djangorestframework==3.11.0
django-jsonfield==1.4.0
gunicorn==20.0.4
idna==2.10
install==1.3.3
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2-binary==2.8.5
pylint==2.5.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
